pkg-export-k8s: Move Cli arg addition to own module
